move indicator specs to the indicators module
