+++ corehq/apps/userreports/specs.py
@@ -1,9 +1,7 @@
-from jsonobject import JsonObject, StringProperty, ListProperty, BooleanProperty, DictProperty
+from jsonobject import JsonObject, StringProperty, ListProperty, DictProperty
 from jsonobject.base import DefaultProperty
-from jsonobject.exceptions import BadValueError
-from corehq.apps.userreports.expressions.getters import (DictGetter, NestedDictGetter, TransformedGetter,
-                                                         transform_date)
-from corehq.apps.userreports.operators import IN_MULTISELECT, EQUAL, OPERATORS
+from corehq.apps.userreports.expressions.getters import (transform_date, getter_from_property_reference)
+from corehq.apps.userreports.operators import OPERATORS
 
 
 def TypeProperty(value):
@@ -15,108 +13,8 @@ def TypeProperty(value):
     return StringProperty(required=True, choices=[value])
 
 
-def DataTypeProperty():
-    """
-    Shortcut for valid data types.
-    """
-    return StringProperty(required=True, choices=['date', 'datetime', 'string', 'integer'])
-
-
-class IndicatorSpecBase(JsonObject):
-    """
-    Base class for indicator specs. All specs (for now) are assumed to have a column_id and
-    a display_name, which defaults to the column_id.
-    """
-    _allow_dynamic_properties = False
-
-    type = StringProperty(required=True)
-
-    column_id = StringProperty(required=True)
-    display_name = StringProperty()
-
-    @classmethod
-    def wrap(cls, obj):
-        wrapped = super(IndicatorSpecBase, cls).wrap(obj)
-        if not wrapped.column_id:
-            raise BadValueError('column_id must not be empty!')
-        if not wrapped.display_name not in obj:
-            wrapped.display_name = wrapped.column_id
-        return wrapped
-
-
-class PropertyReferenceIndicatorSpecBase(IndicatorSpecBase):
-    """
-    Extension of an indicator spec that references a property - either via
-    a property_name or property_path.
-    """
-    property_name = StringProperty()
-    property_path = ListProperty()
-
-    @property
-    def getter(self):
-        return _getter_from_property_reference(self)
-
-
-class BooleanIndicatorSpec(IndicatorSpecBase):
-    type = TypeProperty('boolean')
-    filter = DictProperty(required=True)
-
-
-class RawIndicatorSpec(PropertyReferenceIndicatorSpecBase):
-    type = TypeProperty('raw')
-    datatype = DataTypeProperty()
-    is_nullable = BooleanProperty(default=True)
-    is_primary_key = BooleanProperty(default=False)
-
-    @property
-    def getter(self):
-        transform = _transform_from_datatype(self.datatype)
-        getter = _getter_from_property_reference(self)
-        return TransformedGetter(getter, transform)
-
-
-class ExpressionIndicatorSpec(IndicatorSpecBase):
-    type = TypeProperty('expression')
-    datatype = DataTypeProperty()
-    is_nullable = BooleanProperty(default=True)
-    is_primary_key = BooleanProperty(default=False)
-    expression = DictProperty(required=True)
-
-    @property
-    def parsed_expression(self):
-        from corehq.apps.userreports.expressions.factory import ExpressionFactory
-        transform = _transform_from_datatype(self.datatype)
-        expression = ExpressionFactory.from_spec(self.expression)
-        return TransformedGetter(expression, transform)
-
-
-class ChoiceListIndicatorSpec(PropertyReferenceIndicatorSpecBase):
-    type = TypeProperty('choice_list')
-    choices = ListProperty(required=True)
-    select_style = StringProperty(choices=['single', 'multiple'])
-
-    def get_operator(self):
-        return IN_MULTISELECT if self.select_style == 'multiple' else EQUAL
-
-
 class BaseFilterSpec(JsonObject):
     _allow_dynamic_properties = False
-
-
-def _getter_from_property_reference(spec):
-    if spec.property_name:
-        assert not spec.property_path, \
-            'indicator {} has both a name and path specified! you must only pick one.'.format(spec.property_name)
-        return DictGetter(property_name=spec.property_name)
-    else:
-        assert spec.property_path, spec.property_name
-        return NestedDictGetter(property_path=spec.property_path)
-
-
-def _transform_from_datatype(datatype):
-    return {
-        'date': transform_date
-    }.get(datatype)
 
 
 class BooleanExpressionFilterSpec(BaseFilterSpec):
@@ -134,12 +32,12 @@ class PropertyMatchFilterSpec(BaseFilterSpec):
 
     @property
     def getter(self):
-        return _getter_from_property_reference(self)
+        return getter_from_property_reference(self)
 
 
 class NotFilterSpec(BaseFilterSpec):
     type = TypeProperty('not')
-    filter = DictProperty()  # todo: validators=FilterFactor.validate_spec
+    filter = DictProperty()  # todo: validators=FilterFactory.validate_spec
 
 
 class NamedFilterSpec(BaseFilterSpec):
